Support passing in env slug to execute (#33)
In addition to an env id, we want to support passing through an env slug. Views will use this for local dev - the user may specify an environment in which to test their view by specifying an env slug.

Additionally, passing no env id/slug is valid. This uses the default env.