Forward airplane-internal env vars to activity (#43)
Forwards the env vars from the workflow's process.env into the activity by using `ExecuteOptions` - the workflow and activity run in different Node.js threads and so have different sets of environment variables. The env vars we need are passed in to the workflow when it is started, and so we forward them to the activity too.

Also fixes the run termination signal field names to be capitalized to match the definition in our API, which it is serialized as.