Refactor code layout (#44)
This PR moves most of the codebase into an `internal` folder, which aims to discourage users from directly importing from it. Fields and types that are safe to import are exported from a file in the root. Updates the export format so that built-ins can be imported as `import { query } from 'airplane/sql'`.

cc @ameyakhare