Replace `np` with automatic deployments from GHA (#46)
We were having issues with our previous release process that used `np`. First, it depended on engineers pushing a commit to `main` to change the version field in `package.json`. This does not work with our branch protection rules. Second, it does not support deploying a sub-directory. We build our TS into the `dist` folder and want the built code to be the root of the released package (so that users can import `airplane/sql` which depends on a `sql.js` file in the root of the package). There is a `contents` field from `np` that supports this , but I couldn't get this to work the way we needed.

This PR moves the release process to GHA. It's much more convenient -- just merge a PR that bumps the version in the `package.json` and it will automatically publish. I tested this by releasing `v0.2.0-10`.