Don't retry temporal activities (#73)
By default, temporal activities are retried indefinitely: https://docs.temporal.io/retry-policies/#default-values-for-retry-policy - this is not behavior we'd like since that means multiple child runs could get created in the case of transient temporal failures. In that case, we'd like to fail the workflow entirely - this PR sets the maximum number of attempts in any Node SDK activity to be at most 1 so that retries never occur.