Add dev runtime (#72)
Adds a new runtime type, `dev`, to the Node SDK - previously, `airplane dev` was using the standard runtime methods (successfully); however, this PR formalizes the notion of a new dev runtime with some minor differences from the standard runtime. For example, we don't need to poll for the run to finish - in our local dev api server implementation of `/tasks/execute`, we set the status of the run so there's no need to wait for `/runs/get` to return a result. For now, the `prompt` and `logChunks` methods just reference the standard runtime implementation.