Import nested temporal package (#75)
We are getting warning on dep install that we should be using nested temporal packages:

![image](https://user-images.githubusercontent.com/6363419/179517142-a84ae5e0-b261-451d-ac6d-4283093d4ce2.png)

Switch to using the nested workflow package. 

I'm unsure how to manually test this - any guidance here?

Unrelated, but we need to work on the build system for this repo. It doesn't produce esm modules so it isn't tree shakeable - this is causing consumers (like views) to import all of the code even though we are only using part of the code (e.g. views imports temporal).